separate out, refactor, and document useDataGridSource
