+++ src/components/Grid/useDataGridSource.tsx
@@ -7,10 +7,8 @@ import {
 import {
   FilterColumnsArgs,
   GetColumnForNewFilterArgs,
-  GridColType,
   GridLogicOperator,
   GridSlotProps,
-  GridSortItem,
 } from '@mui/x-data-grid';
 import {
   DataGridProProps as DataGridProps,
@@ -20,14 +18,8 @@ import {
   useGridApiContext,
   useGridApiRef,
 } from '@mui/x-data-grid-pro';
-import { groupBy, Nil, setOf } from '@seedcompany/common';
-import {
-  useDebounceFn,
-  useLatest,
-  useLocalStorageState,
-  useMemoizedFn,
-  useTimeout,
-} from 'ahooks';
+import { groupBy, Nil } from '@seedcompany/common';
+import { useDebounceFn, useMemoizedFn, useTimeout } from 'ahooks';
 import {
   type FieldNode,
   getOperationAST,
@@ -37,18 +29,9 @@ import {
 } from 'graphql';
 import { get, merge, pick, set, uniqBy } from 'lodash';
 import { MutableRefObject, useEffect, useMemo, useRef, useState } from 'react';
-import type { Get, Paths, SetNonNullable } from 'type-fest';
-import { type PaginatedListInput, type SortableListInput } from '~/api';
-import type { Order } from '~/api/schema/schema.graphql';
-import { lowerCase, upperCase } from '~/common';
-import { convertMuiFiltersToApi, FilterShape } from './convertMuiFiltersToApi';
+import type { Get, Paths } from 'type-fest';
 import { useGridFilteredRowCount } from './useGridFilteredRowCount';
-
-type ListInput = SetNonNullable<
-  Required<
-    SortableListInput & PaginatedListInput & { filter?: Record<string, any> }
-  >
->;
+import { ListInput, useViewState, ViewState } from './useViewState';
 
 interface PaginatedListOutput<T> {
   items: readonly T[];
@@ -63,69 +46,58 @@ type PathsMatching<T, List> = {
     : never;
 }[Paths<T>];
 
-const defaultInitialInput = {
-  count: 25,
-  sort: 'id',
-  order: 'ASC' as Order,
-};
 const defaultKeyArgs = ['__typename', 'id'];
-
-const persistColumnTypes = setOf<GridColType>(['singleSelect', 'boolean']);
-
-type StoredViewState = Pick<DataGridProps, 'filterModel'> & {
-  sortModel: [GridSortItem];
-  apiFilterModel?: FilterShape;
-};
-type ViewState = Omit<StoredViewState, 'apiFilterModel'> & {
-  // The sorting state for the first page API query.
-  // It could be the live sorting state, or a stale one,
-  // based on pagination needs.
-  apiSortModel: DataGridProps['sortModel'];
-};
 
 // The built-in NoInfer<T> (TS 5.4+) is an intrinsic that TypeScript doesn't
 // resolve through when accessing properties on constrained generics.
 // This identity-type variant works correctly in that context.
 type NoInfer<T> = [T][T extends any ? 0 : never];
 
-export const useDataGridSource = <
+const emptyList = [] as const;
+
+// ─── useCachedList ──────────────────────────────────────────────────────────
+// Manages all Apollo cache accumulation, page fetching, and row stability.
+// Returns stable rows/total and an imperative onFetchRows for virtual scrolling.
+
+function useCachedList<
   Output extends Record<string, any>,
-  Vars,
-  Input extends Partial<ListInput>,
-  const Path extends PathsMatching<Output, PaginatedListOutput<any>> & string,
-  List extends PaginatedListOutput<any> = Get<Output, Path> extends infer U
-    ? U extends PaginatedListOutput<any>
-      ? U
-      : never
-    : never
+  Vars extends Record<string, any>
 >({
   query,
   variables,
+  variablesWithFilter,
   listAt,
-  initialInput,
-  keyArgs = defaultKeyArgs,
-  apiRef: apiRefInput,
+  keyArgs,
+  hasFilter,
+  input,
+  viewRef,
+  apiRef,
 }: {
   query: DocumentNode<Output, Vars>;
-  variables: NoInfer<Vars & { input?: Input }>;
-  listAt: Path;
-  initialInput?: Partial<Omit<NoInfer<Input>, 'page'>>;
-  keyArgs?: string[];
-  apiRef?: MutableRefObject<GridApiPro>;
-}) => {
-  const initialInputRef = useLatest(initialInput);
-  // eslint-disable-next-line react-hooks/rules-of-hooks -- we'll assume this doesn't change between renders
-  const apiRef = apiRefInput ?? useGridApiRef();
-
-  const filteredRowCount = useGridFilteredRowCount(apiRef);
-
-  const opName = useMemo(
-    () =>
-      query.definitions.find(
-        (d): d is OperationDefinitionNode => d.kind === 'OperationDefinition'
-      )!.name!.value,
-    [query]
-  );
+  variables: Vars;
+  variablesWithFilter: Vars;
+  listAt: string;
+  keyArgs: string[];
+  hasFilter: boolean;
+  input: { count: number } & Record<string, any>;
+  viewRef: MutableRefObject<ViewState>;
+  apiRef: MutableRefObject<GridApiPro>;
+}) {
+  // Helper: extract the list from a query result
+  function listFrom(
+    data: Unmasked<Output> | MaybeMasked<Output>
+  ): PaginatedListOutput<any>;
+  function listFrom(
+    data: Unmasked<Output> | MaybeMasked<Output> | Nil
+  ): PaginatedListOutput<any> | undefined;
+  function listFrom(data: Unmasked<Output> | MaybeMasked<Output> | Nil) {
+    return data ? (get(data, listAt) as PaginatedListOutput<any>) : undefined;
+  }
+  function cacheInfo(data: MaybeMasked<Output> | Nil) {
+    if (!data) return undefined;
+    const list = listFrom(data);
+    return { ...list, isComplete: list.total === list.items.length };
+  }
 
   // Computed once: strip the query down to only keyArgs fields so the
   // cache-accumulation entries written by updateAllPagesQuery stay small.
@@ -144,25 +116,7 @@ export const useDataGridSource = <
     })()
   );
 
-  function listFrom(data: Unmasked<Output> | MaybeMasked<Output>): List;
-  function listFrom(
-    data: Unmasked<Output> | MaybeMasked<Output> | Nil
-  ): List | undefined;
-  function listFrom(data: Unmasked<Output> | MaybeMasked<Output> | Nil) {
-    return data ? (get(data, listAt) as List) : undefined;
-  }
-  function cacheInfo(data: MaybeMasked<Output> | Nil) {
-    if (!data) {
-      return undefined;
-    }
-    const list = listFrom(data);
-    return {
-      ...list,
-      isComplete: list.total === list.items.length,
-    };
-  }
-
-  // Try to pull the complete list from the cache
+  // Try to pull the complete list from the cache.
   // This exists after all pages have been queried with the useQuery hook below.
   // This allows us to do client-side filtering & sorting once we have the complete list.
   const { data: allPagesData, client } = useQuery(query, {
@@ -172,17 +126,13 @@ export const useDataGridSource = <
   const allPages = cacheInfo(allPagesData);
 
   // Add a page to the "all pages" cache entry
-  // This is read back in the first useQuery hook, above.
   const updateAllPagesQuery = (
-    variables: Vars,
+    vars: Vars,
     next: Unmasked<Output>,
     updateTotal: boolean
   ) => {
     client.cache.updateQuery(
-      {
-        query: queryForItemRef.current,
-        variables,
-      },
+      { query: queryForItemRef.current, variables: vars },
       (prev) => {
         const prevList = listFrom(prev);
         const nextList = listFrom(next);
@@ -211,87 +161,6 @@ export const useDataGridSource = <
       }
     );
   };
-
-  // State for current sorting & filtering
-  const [initialSort] = useState((): ViewState['sortModel'] => [
-    {
-      field: initialInput?.sort ?? defaultInitialInput.sort,
-      sort: lowerCase(initialInput?.order ?? defaultInitialInput.order),
-    },
-  ]);
-  const [storedView, setStoredView] = useLocalStorageState<StoredViewState>(
-    `${opName}-data-grid-view`,
-    {
-      defaultValue: () => ({
-        filterModel: { items: [] },
-        apiFilterModel: {},
-        sortModel: initialSort,
-      }),
-    }
-  );
-  const persist = useDebounceFn((next: ViewState) => {
-    const filterModel = {
-      // Strip out filters for columns that shouldn't be persisted
-      items:
-        next.filterModel?.items.filter((item) =>
-          persistColumnTypes.has(apiRef.current.getColumn(item.field).type!)
-        ) ?? [],
-    };
-    setStoredView({
-      sortModel: next.sortModel,
-      filterModel,
-      apiFilterModel: convertMuiFiltersToApi(
-        apiRef.current,
-        filterModel,
-        initialInputRef.current?.filter
-      ),
-    });
-  });
-  const [view, reallySetView] = useState((): ViewState => {
-    const { apiFilterModel: _, ...rest } = storedView!; // not null because we give a default value
-    return {
-      ...rest,
-      apiSortModel: rest.sortModel,
-    };
-  });
-  const setView = (setter: (prev: ViewState) => ViewState) => {
-    reallySetView((prev) => {
-      const next = setter(prev);
-      persist.run(next);
-      return next;
-    });
-  };
-
-  const viewRef = useLatest(view);
-  const hasFilter = !!view.filterModel && view.filterModel.items.length > 0;
-
-  // Convert the view state to the input for the GQL query
-  const input = useMemo(
-    () => ({
-      ...defaultInitialInput,
-      ...initialInputRef.current,
-      count: initialInputRef.current?.count ?? defaultInitialInput.count,
-      ...(view.apiSortModel?.[0] && {
-        sort: view.apiSortModel[0].field,
-        order: upperCase(view.apiSortModel[0].sort!),
-      }),
-      // eslint-disable-next-line no-extra-boolean-cast
-      filter: Boolean(apiRef.current.instanceId)
-        ? convertMuiFiltersToApi(
-            apiRef.current,
-            view.filterModel,
-            initialInputRef.current?.filter
-          )
-        : storedView?.apiFilterModel,
-    }),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [apiRef, initialInputRef, view.apiSortModel, view.filterModel]
-  );
-  const variablesWithFilter = useMemo(() => {
-    const { count, sort, order, ...rest } = input;
-
-    return merge({}, variables, { input: rest });
-  }, [variables, input]);
 
   const addToAllPagesCache = useMemoizedFn((next: MaybeMasked<Output>) => {
     updateAllPagesQuery(variables, next as Unmasked<Output>, !hasFilter);
@@ -348,18 +217,14 @@ export const useDataGridSource = <
     };
   if (freshList != null) prevListRef.current = freshList;
 
-  const rows = list.items ?? emptyList;
+  const rows = list.items;
   const total = list.total && list.total >= 0 ? list.total : undefined;
 
-  // Load additional pages imperatively as needed based on scrolling
-  // This is debounced to mostly to reduce the client side load.
-  // It is theoretical as well that some pages could be scrolled past,
-  // and the debouncing would skip those page requests.
+  // Load additional pages imperatively as needed based on scrolling.
+  // Debounced to reduce client-side load and skip fast-scrolled pages.
   const onFetchRows = useDebounceFn(
     (params: GridFetchRowsParams) => {
-      if (isCacheComplete) {
-        return;
-      }
+      if (isCacheComplete) return;
 
       const { firstRowToRender, lastRowToRender } = params;
       const startPage = Math.ceil((firstRowToRender + 1) / input.count);
@@ -368,7 +233,7 @@ export const useDataGridSource = <
         { length: endPage - startPage + 1 },
         (_, i) => startPage + i
       )
-        // skip the first page always loaded first by useQuery hook
+        // skip the first page always loaded by useQuery above
         .filter((page) => page > 1);
 
       for (const page of pages) {
@@ -385,21 +250,76 @@ export const useDataGridSource = <
               params.sortModel === viewRef.current.sortModel &&
               params.filterModel === viewRef.current.filterModel;
             if (isCurrent) {
-              // Swap in real rows via the recommended process.
               const firstRowToReplace = (page - 1) * input.count;
               const list = listFrom(res.data).items.slice();
               apiRef.current.unstable_replaceRows(firstRowToReplace, list);
             }
-
             // Always try to complete the list in cache.
             addToAllPagesCache(res.data);
           });
       }
     },
-    {
-      wait: 500,
-    }
+    { wait: 500 }
   );
+
+  return { rows, total, loading, isCacheComplete, onFetchRows };
+}
+
+// ─── useDataGridSource ───────────────────────────────────────────────────────
+
+export const useDataGridSource = <
+  Output extends Record<string, any>,
+  Vars,
+  Input extends Partial<ListInput>,
+  const Path extends PathsMatching<Output, PaginatedListOutput<any>> & string
+>({
+  query,
+  variables,
+  listAt,
+  initialInput,
+  keyArgs = defaultKeyArgs,
+  apiRef: apiRefInput,
+}: {
+  query: DocumentNode<Output, Vars>;
+  variables: NoInfer<Vars & { input?: Input }>;
+  listAt: Path;
+  initialInput?: Partial<Omit<NoInfer<Input>, 'page'>>;
+  keyArgs?: string[];
+  apiRef?: MutableRefObject<GridApiPro>;
+}) => {
+  // eslint-disable-next-line react-hooks/rules-of-hooks -- we'll assume this doesn't change between renders
+  const apiRef = apiRefInput ?? useGridApiRef();
+  const filteredRowCount = useGridFilteredRowCount(apiRef);
+
+  const opName = useMemo(
+    () =>
+      query.definitions.find(
+        (d): d is OperationDefinitionNode => d.kind === 'OperationDefinition'
+      )!.name!.value,
+    [query]
+  );
+
+  const {
+    view,
+    setView,
+    viewRef,
+    hasFilter,
+    input,
+    variablesWithFilter,
+    initialSort,
+  } = useViewState({ opName, initialInput, apiRef, variables });
+
+  const { rows, total, loading, isCacheComplete, onFetchRows } = useCachedList({
+    query,
+    variables,
+    variablesWithFilter,
+    listAt,
+    keyArgs,
+    hasFilter,
+    input,
+    viewRef,
+    apiRef,
+  });
 
   const onSortModelChange: DataGridProps['onSortModelChange'] & {} =
     useMemoizedFn((next) => {
@@ -427,6 +347,7 @@ export const useDataGridSource = <
 
       apiRef.current.scrollToIndexes({ rowIndex: 0 });
     });
+
   const onFilterModelChange: DataGridProps['onFilterModelChange'] & {} =
     useMemoizedFn((filterModel) => {
       const next = {
@@ -481,7 +402,7 @@ export const useDataGridSource = <
   return [dataGridProps] as const;
 };
 
-const emptyList = [] as const;
+// ─── Utilities ───────────────────────────────────────────────────────────────
 
 const getFieldPath = (
   x: SelectionSetNode,
